Download and Extract this Zip File
Paste it inside htdocs of XAMPP or run on any other PHP localhost server. For hosting it on a real server you need to change some settings of database in config.php file. One of the important file in the extracted files is arthevents.sql file which needs to be imported in mysql database for the proper functioning of 'arthevents' web application.